python_backend/requirements.txt
Flask
Pillow  # For image processing